requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
requests
openai
python-dotenv